// src/lib/generated/models/transactionResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Finanças Pessoais
 * API de controle de finanças pessoais com categorias, transações, orçamentos e relatórios
 * OpenAPI spec version: 1.0
 */
import type { TransactionResponseDtoType } from './transactionResponseDtoType';
import type { CategoryResponseDto } from './categoryResponseDto';

export interface TransactionResponseDto {
  /** ID da transação */
  id: string;
  /** ID do usuário */
  userId: string;
  /** ID da categoria */
  categoryId: string;
  /** Valor da transação */
  amount: string;
  /**
   * Descrição da transação
   * @nullable
   */
  description: string | null;
  /** Data da transação */
  date: string;
  /** Tipo de transação */
  type: TransactionResponseDtoType;
  /** Categoria da transação */
  category: CategoryResponseDto;
  /** Data de criação */
  createdAt: string;
  /** Data de atualização */
  updatedAt: string;
}